actions: image-partitions: Add option for separate filesystem images
MORE HERE

+++ actions/image_partition_action.go
@@ -182,10 +182,6 @@ import (
 	"encoding/hex"
 	"errors"
 	"fmt"
-	"github.com/docker/go-units"
-	"github.com/freddierice/go-losetup/v2"
-	"github.com/go-debos/fakemachine"
-	"github.com/google/uuid"
 	"log"
 	"os"
 	"os/exec"
@@ -197,6 +193,11 @@ import (
 	"strings"
 	"syscall"
 	"time"
+
+	"github.com/docker/go-units"
+	"github.com/freddierice/go-losetup/v2"
+	"github.com/go-debos/fakemachine"
+	"github.com/google/uuid"
 
 	"github.com/go-debos/debos"
 )
@@ -231,7 +232,10 @@ type imageLocker struct {
 	fd *os.File
 }
 
-func lockImage(context *debos.Context) (*imageLocker, error) {
+func lockImage(context *debos.Context, noop bool) (*imageLocker, error) {
+	if noop {
+		return &imageLocker{}, nil
+	}
 	fd, err := os.Open(context.Image)
 	if err != nil {
 		return nil, err
@@ -244,7 +248,9 @@ func lockImage(context *debos.Context) (*imageLocker, error) {
 }
 
 func (i imageLocker) unlock() {
-	i.fd.Close()
+	if i.fd != nil {
+		i.fd.Close()
+	}
 }
 
 type ImagePartitionAction struct {
@@ -259,6 +265,7 @@ type ImagePartitionAction struct {
 	size             int64
 	loopDev          losetup.Device
 	usingLoop        bool
+	Standalone       bool
 }
 
 func (p *Partition) UnmarshalYAML(unmarshal func(interface{}) error) error {
@@ -324,6 +331,10 @@ func (i *ImagePartitionAction) generateKernelRoot(context *debos.Context) error 
 }
 
 func (i ImagePartitionAction) getPartitionDevice(number int, context debos.Context) string {
+	if i.Standalone {
+		return path.Join(context.Artifactdir, i.ImageName+fmt.Sprintf("%d", number))
+	}
+
 	/* Always look up canonical device as udev might not generate the by-id
 	 * symlinks while there is an flock on /dev/vda */
 	device, _ := filepath.EvalSymlinks(context.Image)
@@ -355,6 +366,12 @@ func (i *ImagePartitionAction) triggerDeviceNodes(context *debos.Context) error 
 
 func (i ImagePartitionAction) PreMachine(context *debos.Context, m *fakemachine.Machine,
 	args *[]string) error {
+
+	if i.Standalone {
+		fmt.Println("PreMachine: running in standalone mode, skipping image creation and partitioning")
+		return nil
+	}
+
 	imagePath := path.Join(context.Artifactdir, i.ImageName)
 	image, err := m.CreateImage(imagePath, i.size)
 	if err != nil {
@@ -467,6 +484,12 @@ func (i ImagePartitionAction) formatPartition(p *Partition, context debos.Contex
 }
 
 func (i *ImagePartitionAction) PreNoMachine(context *debos.Context) error {
+	if i.Standalone {
+		i.usingLoop = false
+		fmt.Println("PreNoMachine: running in standalone mode, skipping image creation and partitioning")
+		return nil
+	}
+
 	imagePath := path.Join(context.Artifactdir, i.ImageName)
 	img, err := os.OpenFile(imagePath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0666)
 	if err != nil {
@@ -511,14 +534,7 @@ func (i *ImagePartitionAction) PreNoMachine(context *debos.Context) error {
 	return nil
 }
 
-func (i ImagePartitionAction) Run(context *debos.Context) error {
-	/* On certain disk device events udev will call the BLKRRPART ioctl to
-	 * re-read the partition table. This will cause the partition devices
-	 * (e.g. vda3) to temporarily disappear while the rescanning happens.
-	 * udev does this while holding an exclusive flock. This means to avoid partition
-	 * devices disappearing while doing operations on them (e.g. formatting
-	 * and mounting) we need to do it while holding an exclusive lock
-	 */
+func (i *ImagePartitionAction) createAndFormatPartitions(context *debos.Context) error {
 	command := []string{"parted", "-s", context.Image, "mklabel", i.PartitionType}
 	if len(i.GptGap) > 0 {
 		command = append(command, i.GptGap)
@@ -627,7 +643,7 @@ func (i ImagePartitionAction) Run(context *debos.Context) error {
 			}
 		}
 
-		lock, err := lockImage(context)
+		lock, err := lockImage(context, i.Standalone)
 		if err != nil {
 			return err
 		}
@@ -642,6 +658,90 @@ func (i ImagePartitionAction) Run(context *debos.Context) error {
 		devicePath := i.getPartitionDevice(p.number, *context)
 		context.ImagePartitions = append(context.ImagePartitions,
 			debos.Partition{Name: p.Name, DevicePath: devicePath})
+	}
+
+	return nil
+}
+
+func (part *Partition) checkSize(i *ImagePartitionAction) (imgSize int64, err error) {
+	var getSizeValueFunc func(size string) (int64, error)
+	if regexp.MustCompile(`^[0-9.]+[kmgtp]ib+$`).MatchString(strings.ToLower(i.ImageSize)) {
+		getSizeValueFunc = units.RAMInBytes
+	} else {
+		getSizeValueFunc = units.FromHumanSize
+	}
+
+	start, err := getSizeValueFunc(part.Start)
+	if err != nil {
+		return 0, fmt.Errorf("failed to parse partition start size: %w", err)
+	}
+	if i.Standalone && start != 0 {
+		return 0, fmt.Errorf("invalid partition start size: %s, when in standalone mode start should be 0", part.Start)
+	} else if !i.Standalone && start < 0 {
+		return 0, fmt.Errorf("invalid partition start size: %s, when not in standalone mode start should be greater than or equal to 0", part.Start)
+	}
+
+	end, err := getSizeValueFunc(part.End)
+	if err != nil {
+		return 0, fmt.Errorf("failed to parse partition end size: %w", err)
+	}
+	if end <= 0 {
+		return 0, fmt.Errorf("invalid partition end size: %s, end should be greater than 0", part.End)
+	}
+
+	return end - start, nil
+}
+
+func (i *ImagePartitionAction) createStandalonePartitions(context *debos.Context) error {
+	for idx := range i.Partitions {
+		current_partition := &i.Partitions[idx]
+
+		imagePath := path.Join(context.Artifactdir, i.ImageName+"-"+current_partition.Name)
+		image, err := os.OpenFile(imagePath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0666)
+
+		if err != nil {
+			return fmt.Errorf("couldn't open partition file: %w", err)
+		}
+
+		size, err := current_partition.checkSize(i)
+		if err != nil {
+			return fmt.Errorf("invalid partition size: %w", err)
+		}
+
+		err = image.Truncate(size)
+		if err != nil {
+			return fmt.Errorf("couldn't resize partition file: %w", err)
+		}
+
+		image.Close()
+		err = i.formatPartition(current_partition, *context)
+		if err != nil {
+			return err
+		}
+
+		devicePath := imagePath
+		context.ImagePartitions = append(context.ImagePartitions,
+			debos.Partition{Name: i.Partitions[idx].Name, DevicePath: devicePath})
+	}
+
+	return nil
+}
+
+func (i ImagePartitionAction) Run(context *debos.Context) error {
+	/* On certain disk device events udev will call the BLKRRPART ioctl to
+	 * re-read the partition table. This will cause the partition devices
+	 * (e.g. vda3) to temporarily disappear while the rescanning happens.
+	 * udev does this while holding an exclusive flock. This means to avoid partition
+	 * devices disappearing while doing operations on them (e.g. formatting
+	 * and mounting) we need to do it while holding an exclusive lock
+	 */
+
+	if i.Standalone {
+		fmt.Println("Running in standalone mode, skipping image creation and partitioning")
+		i.createStandalonePartitions(context)
+	} else {
+		fmt.Println("Running in non-standalone mode, creating and partitioning image")
+		i.createAndFormatPartitions(context)
 	}
 
 	context.ImageMntDir = path.Join(context.Scratchdir, "mnt")
@@ -665,7 +765,7 @@ func (i ImagePartitionAction) Run(context *debos.Context) error {
 		return strings.Count(mntA, "/") < strings.Count(mntB, "/")
 	})
 
-	lock, err := lockImage(context)
+	lock, err := lockImage(context, i.Standalone)
 	if err != nil {
 		return err
 	}
@@ -682,7 +782,7 @@ func (i ImagePartitionAction) Run(context *debos.Context) error {
 		case "fat", "fat12", "fat16", "fat32", "msdos":
 			fsType = "vfat"
 		}
-		err = syscall.Mount(dev, mntpath, fsType, 0, "")
+		err = debos.Command{}.Run("mount", "mount", dev, mntpath, "-t", fsType)
 		if err != nil {
 			return fmt.Errorf("%s mount failed: %w", m.part.Name, err)
 		}
@@ -697,6 +797,11 @@ func (i ImagePartitionAction) Run(context *debos.Context) error {
 	err = i.generateKernelRoot(context)
 	if err != nil {
 		return err
+	}
+
+	if i.Standalone {
+		fmt.Println("Standalone mode: skipping triggering device nodes")
+		return nil
 	}
 
 	/* Now that all partitions are created (re)trigger all udev events for
@@ -760,9 +865,20 @@ func (i ImagePartitionAction) PostMachineCleanup(context *debos.Context) error {
 	image := path.Join(context.Artifactdir, i.ImageName)
 	/* Remove the image in case of any action failure */
 	if context.State != debos.Success {
-		if _, err := os.Stat(image); !os.IsNotExist(err) {
-			if err = os.Remove(image); err != nil {
-				return err
+		if i.Standalone {
+			for partion := range context.ImagePartitions {
+				image := path.Join(context.Artifactdir, i.ImageName+"-"+context.ImagePartitions[partion].Name)
+				if _, err := os.Stat(image); !os.IsNotExist(err) {
+					if err = os.Remove(image); err != nil {
+						return err
+					}
+				}
+			}
+		} else {
+			if _, err := os.Stat(image); !os.IsNotExist(err) {
+				if err = os.Remove(image); err != nil {
+					return err
+				}
 			}
 		}
 	}
@@ -863,7 +979,7 @@ func (i *ImagePartitionAction) Verify(_ *debos.Context) error {
 			}
 		}
 
-		if i.PartitionType != "gpt" && p.PartLabel != "" {
+		if i.PartitionType != "gpt" && p.PartLabel != "" && !i.Standalone {
 			return fmt.Errorf("can only set partition partlabel on GPT filesystem")
 		}
 
@@ -879,7 +995,7 @@ func (i *ImagePartitionAction) Verify(_ *debos.Context) error {
 			}
 		}
 
-		if p.PartType != "" {
+		if p.PartType != "" && !i.Standalone {
 			var partTypeLen int
 			switch i.PartitionType {
 			case "gpt":
@@ -971,21 +1087,25 @@ func (i *ImagePartitionAction) Verify(_ *debos.Context) error {
 		}
 	}
 
-	// Calculate the size based on the unit (binary or decimal)
-	// binary units are multiples of 1024 - KiB, MiB, GiB, TiB, PiB
-	// decimal units are multiples of 1000 - KB, MB, GB, TB, PB
-	var getSizeValueFunc func(size string) (int64, error)
-	if regexp.MustCompile(`^[0-9.]+[kmgtp]ib+$`).MatchString(strings.ToLower(i.ImageSize)) {
-		getSizeValueFunc = units.RAMInBytes
-	} else {
-		getSizeValueFunc = units.FromHumanSize
+	//Not calculating the image size in standalone mode as it is not needed and can be error-prone due to possible overlaps of partitions
+	if !i.Standalone {
+		// Calculate the size based on the unit (binary or decimal)
+		// binary units are multiples of 1024 - KiB, MiB, GiB, TiB, PiB
+		// decimal units are multiples of 1000 - KB, MB, GB, TB, PB
+		var getSizeValueFunc func(size string) (int64, error)
+		if regexp.MustCompile(`^[0-9.]+[kmgtp]ib+$`).MatchString(strings.ToLower(i.ImageSize)) {
+			getSizeValueFunc = units.RAMInBytes
+		} else {
+			getSizeValueFunc = units.FromHumanSize
+		}
+
+		size, err := getSizeValueFunc(i.ImageSize)
+		if err != nil {
+			return fmt.Errorf("failed to parse image size: %s", i.ImageSize)
+		}
+
+		i.size = size
 	}
 
-	size, err := getSizeValueFunc(i.ImageSize)
-	if err != nil {
-		return fmt.Errorf("failed to parse image size: %s", i.ImageSize)
-	}
-
-	i.size = size
 	return nil
 }
